issues with enitity, component, design

+++ src/grid.rs
@@ -37,8 +37,8 @@ pub struct GameInstance {
     pub x_size: usize,
     pub y_size: usize,
     pub grid: Vec<Vec<Tile>>,
-    pub grid_size: [f64; 2], //Window size should use x_size and y_size
-    pub window_size: [f64; 2],
+    pub grid_size: [f32; 2], //Window size should use x_size and y_size
+    pub window_size: [f32; 2],
     pub flag_count: i32, //Variable for storing how many flags are left mutable
     pub bomb_count: i32, //Variable for storing how many bombs are planted - STATIC
     pub game_state: i32,
@@ -92,15 +92,15 @@ pub fn create_grid(x_size: usize, y_size: usize) -> Vec<Vec<Tile>> {
     return vec;
 }
 //Function that plants bombs over the grid
-pub fn plant_bombs(game: &mut GameInstance, coords: [f64; 2]){
+pub fn plant_bombs(game: &mut GameInstance, coords: [f32; 2]){
     //Get tile for first tile selected
     let tile_coords = find_tile(coords, &game);
     let row_len = game.grid.len();
     let col_len = game.grid[0].len();
     let tile_count = game.grid.len() * game.grid[0].len();//Get the total number of tiles
     let mut rng = thread_rng();
-    let y: f64 = rng.gen_range(0.75, 1.25);
-    let mut bomb_count = ((tile_count/4) as f64 * y).round() as i32;
+    let y: f32 = rng.gen_range(0.75, 1.25);
+    let mut bomb_count = ((tile_count/4) as f32 * y).round() as i32;
     game.bomb_count = bomb_count;
     game.flag_count = bomb_count;
     println!("Bomb Count {}", bomb_count);
@@ -110,9 +110,9 @@ pub fn plant_bombs(game: &mut GameInstance, coords: [f64; 2]){
         // if (row != tile_coords[1] && col != tile_coords[0]) ||
         // (row != tile_coords[1]-1 && col != tile_coords[0]-1) || 
         // (row != tile_coords[1]-1 && col != tile_coords[0]-1) || 
-        let r = row as f64;
-        let c = col as f64;
-        if !(r as f64 >= tile_coords[1]-1.0 && c >= tile_coords[0]-1.0 && r <= tile_coords[1]+1.0 && c <= tile_coords[0]+1.0) {
+        let r = row as f32;
+        let c = col as f32;
+        if !(r as f32 >= tile_coords[1]-1.0 && c >= tile_coords[0]-1.0 && r <= tile_coords[1]+1.0 && c <= tile_coords[0]+1.0) {
             if game.grid[row as usize][col as usize].value != -1 {
                 game.grid[row as usize][col as usize] = Tile::bomb();
                 bomb_count -= 1;
@@ -141,8 +141,8 @@ pub fn debug_map(vec: &Vec<Vec<Tile>>, hidden: bool){
     }
 }
 //Converts mouse coordinates to tile
-fn find_tile(coords: [f64; 2], game: &GameInstance) -> [f64; 2]{
-    let tile_size: [f64; 2] = [game.grid_size[0]/game.x_size as f64, game.grid_size[1]/game.y_size as f64];
+fn find_tile(coords: [f32; 2], game: &GameInstance) -> [f32; 2]{
+    let tile_size: [f32; 2] = [game.grid_size[0]/game.x_size as f32, game.grid_size[1]/game.y_size as f32];
     let x = coords[0] / tile_size[0];
     let y = coords[1] / tile_size[1];
     //println!("X: {}, Y: {}", x, y);
@@ -297,7 +297,7 @@ fn show_empty_edges(game: &mut GameInstance, shown_tiles: Vec<[i32; 2]>){
     }
 }
 
-pub fn left_click(coords: [f64; 2], game: &mut GameInstance){
+pub fn left_click(coords: [f32; 2], game: &mut GameInstance){
     if click_on_grid(&coords, &game) {
         let tile_coords = find_tile(coords, &game);
         if game.grid[tile_coords[1] as usize][tile_coords[0] as usize].value == -1 {
@@ -316,7 +316,7 @@ pub fn left_click(coords: [f64; 2], game: &mut GameInstance){
     }
 }
 
-pub fn right_click(coords: [f64; 2], game: &mut GameInstance){
+pub fn right_click(coords: [f32; 2], game: &mut GameInstance){
     if click_on_grid(&coords, &game) {
         let tile_coords = find_tile(coords, &game);
         if game.grid[tile_coords[1] as usize][tile_coords[0] as usize].hidden == true {
@@ -331,7 +331,7 @@ pub fn right_click(coords: [f64; 2], game: &mut GameInstance){
     }
 }
 
-pub fn click_on_grid(coords: &[f64; 2], game: &GameInstance) -> bool{
+pub fn click_on_grid(coords: &[f32; 2], game: &GameInstance) -> bool{
     if coords[0] > game.grid_size[1] || coords[1] > game.grid_size[0] {
         false
     } else {
